requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
pymupdf
pytesseract
Pillow
sentence-transformers
groq
python-dotenv